Fix unwrap panic during install with unsupported modules (#298)
Description
===========

The install2 `install` function panics when one or more modules
are selected, which are not compatible with the selected unity
version to install. The component enum values are not save to
use over all versions of unity. This can only be determined at
runtime. The old implementation did a simple unwrap and paniced.
I rewrote the logic to yield a `Result<HashSet<Component>>` instead.