Fix error print in --help messages. (#314)
This patch fixes a small output bug. When
calling the help page, the command would
print an error message which comes from the
default rust implementation when main returns an
error result. This patch simply calls the failing
method of `structopt` `from_args` which will panic.

It also enables color output for the cli help.