Fix iOSSupport module installation destination on linux
Description
===========

The `iOSSupport` module is the only package on `macOS` which is archived
based on a different base directory than all other packages. This is sadly
also true for the linux version of the module. So I adjusted the installation
destination when the package destination contains `iOSSupport` in the path.