Remove dependency to cli-core
Description
===========

`cli-core` was a crate a prematurely created and published. I was
in the mood of writing other internal tools and liked the idea
to use one common way of setting up the commandline interfact.
The problem is that this all builds on top of docopt and
works around some of the shortcommings. Since I want to move to
clap v3 (with the middle solution structopt) I decided to get rid
of this dependency which was the last straw which pulled on docopt.