Improve unity version error handling (#350)
Description
===========

The unity version module already had a custom error
implementation based on the `error-chain` cask.
I reimplemented the error based on `thiserror` and cleaned up
some cases. I'm not 100% happy about the result but would revisit
this again at a later stage. I'm not super fond of the introduction
of anyhow as a catch all item.