Improve unity project version detection
The last patch to `uvm_core` added the ability to parse
Unity version with the added revision hash. This patch changes
the `dectect_project_version` to favior the `m_EditorVersionWithRevision` in
the ProjectSettings.txt file. This will return a version with the added
revision hash. I moved some code around to clean this up a little bit.