Improve install cache
Description
===========

The installer cache worked fine and has functionally no issues. The reason
I was compelled to patch it is that Unity sometimes creates internal fix
releases with a different revision hash. This means that when the actual
fix gets rolled out, uvm might think that the installer for the version
is already downloaded.
This is a very simple patch to make sure that we use both user version + revision hash
as the cache key for the installer and lock files.

While touching it I moved the cache folder to a new home. I use the
github domain for the cache dir instead of `Wooga/UnityVersionManager`.
I also decided to keep the temp download files in a seperate directory.