Remove LinuxMono from list of Unity built-in components (#377)
The `LinuxMono` unity component is listed as one of the default 'built-in' linux components. While unity work as the linux-mono component is already installed, this generates an inconsistency with other OSes that doesn't list their [os]-mono packages, which behave in the same way. 
This PR removes the LinuxMono from the list built-in linux components to keep consistency.