Add new crate to connect to unity live platform service
Description
===========

This is a first patch to introduce a new crate to connect to
the live platform service. This service is a graphql service which
allows a handfull of operations around unity releases etc. This service
is used directly by UnityHub to query latest releases and to fetch
information for older releases. This used to be semi-hardcoded in Unity Hub.
The first API allows to query for Unity versions. Others will follow soon.

Changes
=======

* ![ADD] `uvm_live_platform` crate to connect to `live-platform` service