Add new sub crate unity-version (#390)
Description
===========

This is a new implementation of the Unity version struct and parsing
methods to represent a Unity version inside uvm. It will be used
with future patches. The old representation tried to combine a normal unity version
with an optional revision hash. I constructed this as a seperate type. Conversion
methods are not yet available because I need to see how I will actually use both
types.