Fix locking logic
Description
===========

The locking logic had a slight bug as it treats all types of errors
equally. But we only need to check for already locked type of an error.
I think in praxis this would have not come up since the next call to
wait_lock would failed as well. But the intention is cleaner.