Add a new optional check if the already installed version is architecture compatible (#402)
Description
===========

To support a clean upgrade path we need to make sure we auto cleanup
intel Unity installation of arm based macs.

I added a helper function which determines if an already installed
Unity version is binary compatible with the current system. This check
can be enabled with a system environment variable `UVM_ARCHITECTURE_CHECK_ENABLED`.
The whole function is only implemented for macOS targets with arm architecture.
If enabled the installer will delete the installation and reinstall.