Add support to list and query XLTS releases (#403)
Description
===========

Unity has a new LTS version scheme called XLTS which means
extended LTS for enterprise customer. By default these can't be listed
with the graphql endpoints. One needs to provide an entitlements list to
include this versions in the query. The documentation describes them as
a filter but it acts more like an inclusion boolean.
I also added the second valid entitlement U7_Alpha.

I also changed the builder pattern and usage so the default is more a
default and not opinionated.