Restructure codebase and move old comamnds to legacy dir (#408)
Description
============

This is yet another patch to restructrure the tool.
I need to get rid of some commands on the main workspace because they
refer to uvm_core. But I don't want to delete the code just yet. So I
moved some crates into a new workspace and improved the folder structure
a bit.