Inline uvm-uninstall command to uvm crate (#409)
Description
===========

This is a bigger patch to inline uvm-uninstall.
This one is not simply a move to the uvm crate but also a refactor of
both the cli interface and the logic.

The old tool took named options like `--ios` '--android` etc to
determine the modules to uninstall. But this is no longer supported same
as in install2. So instead one needs to specify the module ids via
`--module` option. The uninstall impl has changed with a basic directory
delete. Since the tool no longer has a static map from module id to
installation location it tries to figure this out via the modules json.
Only modules which are located inside the installation directory can be
unisntalled like this. Also modules that point to the root directory
also can't be uninstalled. This will happen mostly on windows because
the installers for the platform supports get the path to the
installation passed not the actual directory where the installer will
install the module.

There is also a second smaller refactor. The uninstall logic nees to
modify the modules.json file of the installation. This feature was only
present as a function in the install crate. Now it is moved to the unity
hub crate guarded behind a feature flag.