Fix Linux binary version parsing
Description
===========

I added a regression in the Linux ports of uvm where the version
of Unity could not be detected anymore when the logic did a fallback
to reading the editor version from the binary. The `FromStr`
implementation of `Version` coming from `upm_core` used to be very
forgiving what strings got passed. We used to pass the whole outputs
from `strings UNITY_INSTALL_PATH/Editor/Unity`. This is no longer
supported and a helper function should be called. I took the time to
improve this as well since newer Unity version seem to mention other
versions in their internal helper string quite a lot. So the logic works
with a multi tier search to priotize certain formats over others.
The actual version printed in unity is always the one with the Hash for
instance.