Add caching for live platform api (#432)
Description
===========

The requests to the graphql endpoint can take quite long depending
on the query. Because of how some of the commands filter the output
and how often the responses actually change it makes no sense to call
the service every single time. So this patch adds a caching service
which caches responses per request level.