Fix uvm modules command output
Description
===========

The modules command was using the current system platform and
architecture. Without any extra logic older versions of Unity would
simply show "Failed to load version". Instead of this unhelpful error
the new logic is to load the data for all platforms and list
the information accordingly. The command also has a new flag `--current`
to filter out modules which are not compatible with the current system /
arch. If the list ends up in an empty list an error message will show
so.