Improve GC run (#446)
Description
===========

Improve when GC runs. I moved from all commands to a subset.
Not all uvm commands generate garbage and running gc on every
command might be a bit overload. I also added a way to run the gc
command with an optional version parameter to cleanup cached items for
a specific unity version. I opted for a full version.